# src/apps/apis/communities-api/routes/DeleteCommunityBanRoute.ts
import CommunityMemberUnbanner from '@app/contexts/communities/application/ban-member/CommunityMemberUnbanner';
import { CommunityMemberUnbanMessage } from '@app/contexts/communities/application/ban-member/messages/CommunityMemberUnbanMessage';
import { HttpRouteStatusEnum } from '@haskou/ddd-kernel/contracts/ui';
import { Request, Response } from 'express';
import {
  Body,
  Delete,
  JsonController,
  Param,
  Req,
  Res,
} from 'routing-controllers';

import { DeleteCommunityBanBody } from '../bodies/DeleteCommunityBanBody';
import { CommunityViewModel } from '../view-model/CommunityViewModel';
import { CommunityRouteSupport } from './CommunityRouteSupport';

@JsonController('/communities')
export class DeleteCommunityBanRoute extends CommunityRouteSupport {
  private readonly unbanner = this.get<CommunityMemberUnbanner>(
    CommunityMemberUnbanner,
  );

  @Delete('/:communityId/bans/:identityId')
  public async unbanMember(
    @Param('communityId') communityId: string,
    @Param('identityId') identityId: string,
    @Body() body: DeleteCommunityBanBody,
    @Req() request: Request,
    @Res() response: Response,
  ): Promise<Response> {
    const actorIdentityId = await this.authenticate(request);
    const community = await this.unbanner.unban(
      new CommunityMemberUnbanMessage(
        communityId,
        actorIdentityId.valueOf(),
        identityId,
        body.moderationLog,
      ),
    );

    return response
      .status(HttpRouteStatusEnum.OK)
      .send(new CommunityViewModel(community).toResource());
  }
}