# src/apps/apis/communities-api/swagger.yaml
openapi: 3.0.3
info:
  title: Pigeon-swarm Communities API
  version: 1.0.0
  description: Endpoints for private/public communities, members, text channels, and channel messages.
...
paths:
  /communities/:
    get:
      tags:
        - communities
      summary: List communities for the authenticated member
      parameters:
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      responses:
        '200':
          description: Communities where the authenticated identity is a member.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunitiesResource'
    post:
      tags:
        - communities
      summary: Create a community
      description: The authenticated identity becomes the owner and first member. visibility defaults to private and cannot be changed later. Private communities require encrypted channel message payloads. Public communities require plaintext channel message payloads so messages can be indexed and searched.
      parameters:
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommunityBody'
      responses:
        '200':
          description: Community created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
...
  /communities/membership-requests/{requestId}:
    patch:
      tags:
        - communities
      summary: Accept or decline a community invitation or join request
      description: Invited identities can accept or decline invitations. Community owners or members with approve/reject member permissions can accept or decline join requests. Accepted requests add the target identity to memberIds.
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchCommunityMembershipRequestBody'
      responses:
        '200':
          description: Membership request updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityMembershipRequestResource'
...
  /communities/{communityId}/bans/{identityId}:
    delete:
      tags:
        - communities
      summary: Remove a community ban
      description: Requires ban_members.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - in: path
          name: identityId
          required: true
          schema:
            type: string
          description: URL-encoded identity id.
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteCommunityBanBody'
      responses:
        '200':
          description: Ban removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
...
  /communities/{communityId}/roles/{roleId}:
    patch:
      tags:
        - communities
      summary: Update a community role
      description: Only the owner or members with manage_roles can update roles.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - in: path
          name: roleId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunityRoleBody'
      responses:
        '200':
          description: Role updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
    delete:
      tags:
        - communities
      summary: Delete a community role
      description: Only the owner or members with manage_roles can delete custom roles. Built-in roles cannot be deleted.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - in: path
          name: roleId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteCommunityRoleBody'
      responses:
        '200':
          description: Role deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
...
  /communities/{communityId}/channels/{channelId}:
    patch:
      tags:
        - communities
      summary: Rename a private community text channel
      description: Requires manage_channels.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - in: path
          name: channelId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchCommunityChannelBody'
      responses:
        '200':
          description: Channel renamed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
        '409':
          description: Authenticated identity does not have manage_channels or the channel does not exist.
    delete:
      tags:
        - communities
      summary: Delete a private community channel
      description: Requires manage_channels. Deleting a text channel also deletes its stored channel messages.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - in: path
          name: channelId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteCommunityChannelBody'
      responses:
        '200':
          description: Channel deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityResource'
        '409':
          description: Authenticated identity does not have manage_channels or the channel does not exist.
...
  /communities/{communityId}/channels/{channelId}/messages/{messageId}:
    put:
      tags:
        - communities
      summary: Edit a community text channel message
      description: Only the original author can edit a channel message. Private communities require encryptedPayload and public communities require plaintextPayload. The edit carries a signed `mutation` (higher sequence on the same record id) and the backend announces the change through realtime and pubsub events.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - in: path
          name: channelId
          required: true
          schema:
            type: string
        - in: path
          name: messageId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCommunityChannelMessageBody'
      responses:
        '200':
          description: Message edited.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunityChannelMessageResource'
        '409':
          description: Authenticated identity is not the author, the channel does not exist, or the public mutation is invalid.
    delete:
      tags:
        - communities
      summary: Delete an encrypted private community text channel message
      description: Only community members can delete channel messages. The deletion is announced through realtime and pubsub events.
      parameters:
        - $ref: '#/components/parameters/CommunityId'
        - in: path
          name: channelId
          required: true
          schema:
            type: string
        - in: path
          name: messageId
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XIdentityId'
        - $ref: '#/components/parameters/XTimestamp'
        - $ref: '#/components/parameters/XSignature'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteCommunityChannelMessageBody'
      responses:
        '200':
          description: Message deleted locally.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedCommunityChannelMessageResource'
        '409':
          description: Authenticated identity is not a member, the channel does not exist, or the public mutation is invalid.
...
components:
  parameters:
    CommunityId:
      in: path
      name: communityId
      required: true
      schema:
        type: string
    XIdentityId:
      in: header
      name: X-Identity-Id
      required: true
      schema:
        type: string
    XTimestamp:
      in: header
      name: X-Timestamp
      required: true
      schema:
        type: string
    XSignature:
      in: header
      name: X-Signature
      required: true
      schema:
        type: string
  schemas:
    CommunityTextChannelResource:
      type: object
      required:
        - id
        - name
        - type
        - createdAt
        - permissions
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - text
        createdAt:
          type: integer
        permissions:
          $ref: '#/components/schemas/CommunityChannelPermissionsResource'
        threads:
          type: array
          description: Recent active threads for this text channel, newest activity first.
          items:
            $ref: '#/components/schemas/CommunityChannelThreadSummaryResource'
...
    CommunityVoiceChannelResource:
      type: object
      required:
        - id
        - name
        - type
        - createdAt
        - connectedIdentityIds
        - permissions
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - voice
        createdAt:
          type: integer
        permissions:
          $ref: '#/components/schemas/CommunityChannelPermissionsResource'
        connectedIdentityIds:
          type: array
          description: Unique identities joined to active calls for this voice channel with at least one connected participant lease for the same call and identity. Historical joined participants without a connected lease are excluded.
          items:
            type: string
    CommunityResource:
      type: object
      required:
        - id
        - networkId
        - ownerIdentityId
        - name
        - description
        - memberIds
        - bannedMemberIds
        - memberRoles
        - roles
        - textChannels
        - voiceChannels
        - visibility
        - discoverable
        - autoJoinEnabled
        - createdAt
      properties:
        id:
          type: string
        networkId:
          type: string
        ownerIdentityId:
          type: string
        name:
          type: string
        description:
          type: string
        discoverable:
          type: boolean
          description: Whether this private community appears in discovery search results.
        autoJoinEnabled:
          type: boolean
          description: Whether non-banned identities can join without owner approval.
        avatar:
          type: string
          description: Public IPFS CID for the community avatar/icon.
        banner:
          type: string
          description: Public IPFS CID for the community banner.
        memberIds:
          type: array
          items:
            type: string
        bannedMemberIds:
          type: array
          items:
            type: string
        memberRoles:
          type: array
          items:
            $ref: '#/components/schemas/CommunityMemberRolesResource'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/CommunityRoleResource'
        textChannels:
          type: array
          items:
            $ref: '#/components/schemas/CommunityTextChannelResource'
        voiceChannels:
          type: array
          items:
            $ref: '#/components/schemas/CommunityVoiceChannelResource'
        visibility:
          type: string
          enum:
            - private
            - public
        createdAt:
          type: integer
    CommunitiesResource:
      type: object
      required:
        - communities
      properties:
        communities:
          type: array
          items:
            $ref: '#/components/schemas/CommunityResource'
...
    PatchCommunityMembershipRequestBody:
      type: object
      description: The mutation signs the stored request document (same fields, new status and updatedAt) in store requests, recordId = request id, sequence above the previous one. Signer is the invitee (accept/decline an invitation), the creator (decline an invitation), a moderator or the owner (resolve a request) or the requester (decline own request).
      required:
        - status
        - updatedAt
        - mutation
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        updatedAt:
          type: integer
          minimum: 0
        mutation:
          $ref: '#/components/schemas/SignedPublicMutation'
        status:
          type: string
          enum:
            - accepted
            - declined
    CommunityMembershipRequestResource:
      type: object
      required:
        - id
        - communityId
        - creatorIdentityId
        - identityId
        - type
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        communityId:
          type: string
        creatorIdentityId:
          type: string
        identityId:
          type: string
        type:
          type: string
          enum:
            - invitation
            - request
        status:
          type: string
          enum:
            - accepted
            - declined
            - pending
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    CommunityRoleResource:
      type: object
      required:
        - id
        - name
        - permissions
        - builtIn
      properties:
        id:
          type: string
        name:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/CommunityPermission'
        builtIn:
          type: boolean
    CommunityMemberRolesResource:
      type: object
      required:
        - identityId
        - roleIds
      properties:
        identityId:
          type: string
        roleIds:
          type: array
          items:
            type: string
...
    CommunityChannelMessageResource:
      type: object
      required:
        - id
        - communityId
        - channelId
        - authorIdentityId
        - mentions
        - reactions
        - type
        - createdAt
      properties:
        id:
          type: string
        communityId:
          type: string
        channelId:
          type: string
        authorIdentityId:
          type: string
        encryptedPayload:
          type: string
          description: Present for sent private community messages. Omitted for public messages and poll timeline items.
        plaintextPayload:
          type: string
          description: Present for sent public community messages. Omitted for private messages and poll timeline items.
        mentions:
          type: array
          description: Visible mention metadata used for permission checks and notifications. Content remains encrypted.
          items:
            $ref: '#/components/schemas/CommunityMessageMentionResource'
        type:
          type: string
          enum:
            - poll
            - sent
        pollId:
          type: string
          description: Present for poll timeline items. The value matches id.
        replyToMessageId:
          type: string
          description: Optional id of the message this channel message replies to.
        createdAt:
          type: integer
        editedAt:
          type: integer
        reactions:
          type: array
          description: OrbitDB replicated reactions synchronized by community events.
          items:
            $ref: '#/components/schemas/CommunityChannelMessageReactionResource'
    CommunityChannelMessageReactionResource:
      type: object
      required:
        - authorIdentityId
        - createdAt
        - emoji
      properties:
        authorIdentityId:
          type: string
        createdAt:
          type: integer
        emoji:
          type: string
...
    DeletedCommunityChannelMessageResource:
      type: object
      required:
        - id
        - communityId
        - channelId
        - targetMessageId
        - deletedByIdentityId
        - type
      properties:
        id:
          type: string
        communityId:
          type: string
        channelId:
          type: string
        targetMessageId:
          type: string
        deletedByIdentityId:
          type: string
        type:
          type: string
          enum:
            - deleted
    PostCommunityBody:
      type: object
      required:
        - networkId
        - name
        - description
      properties:
        networkId:
          type: string
        name:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 500
        avatar:
          type: string
          description: Optional public IPFS CID for the community avatar/icon.
        banner:
          type: string
          description: Optional public IPFS CID for the community banner.
        discoverable:
          type: boolean
          default: true
          description: Set false to keep this community out of discovery search results.
        autoJoinEnabled:
          type: boolean
          default: false
          description: Set true to let any non-banned identity join without owner approval.
        visibility:
          type: string
          default: private
          description: Select private for encrypted channel messages or public for plaintext searchable channel messages. This value is immutable after creation.
          enum:
            - private
            - public
...
    DeleteCommunityBanBody:
      type: object
      required:
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
    DeleteCommunityRoleBody:
      type: object
      required:
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
    DeleteCommunityChannelBody:
      type: object
      required:
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
    CommunityModerationLogBody:
      type: object
      description: Client-signed moderation log entry. The mutation signs {action, actorIdentityId, communityId, createdAt, details, id, scopeType 'community_moderation_log', target {id, type}} in store moderationLogs, recordId = id, sequence 0, signer = the acting moderator. id is the first 24 hex chars of sha256(JSON.stringify([communityId, actorIdentityId, action, target.type, target.id, createdAt])). Channel and role creation use createdAt as the input of the derived channel/role id.
      required:
        - createdAt
        - mutation
      properties:
        createdAt:
          type: integer
          minimum: 0
        mutation:
          $ref: '#/components/schemas/SignedPublicMutation'
...
    PatchCommunityChannelBody:
      type: object
      required:
        - name
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        name:
          type: string
          maxLength: 80
    CommunityRoleBody:
      type: object
      required:
        - name
        - permissions
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        name:
          type: string
          maxLength: 50
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/CommunityPermission'
...
    CommunityPermission:
      type: string
      enum:
        - approve_members
        - attach_files
        - ban_members
        - connect_voice
        - create_invites
        - create_polls
        - embed_links
        - manage_channels
        - manage_members
        - manage_messages
        - manage_roles
        - mention_everyone
        - mention_here
        - mention_roles
        - reject_members
        - send_messages
        - send_stickers
        - view_channels
...
    CommunityMessageMentionResource:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - everyone
            - here
            - identity
            - role
        targetId:
          type: string
          description: Identity id or role id for identity/role mentions.
    DeleteCommunityChannelMessageBody:
      type: object
      required:
        - mutation
        - moderationLog
      properties:
        moderationLog:
          $ref: '#/components/schemas/CommunityModerationLogBody'
        mutation:
          $ref: '#/components/schemas/SignedPublicMutation'
    PutCommunityChannelMessageBody:
      type: object
      required:
        - createdAt
        - mutation
      properties:
        createdAt:
          type: integer
          description: Edition timestamp used in the domain signature and returned as editedAt.
        encryptedPayload:
          type: string
          description: Required when editing private community messages. Updated client-encrypted payload stored as opaque text.
        plaintextPayload:
          type: string
          description: Required when editing public community messages. Updated plain text payload stored for indexing/search.
        mutation:
          $ref: '#/components/schemas/SignedPublicMutation'
        mentions:
          type: array
          items:
            $ref: '#/components/schemas/CommunityMessageMentionResource'
    SignedPublicMutation:
      type: object
      description: Client-signed intent for a replicated public record. The node holds no user private keys, so the UI signs it. payloadDigest commits to the record without its proof; author.deviceCredential must be an authorized device of author.identityId in its current device authorization head. The winner of concurrent mutations is the higher sequence, then the lower digest; no wall clock is used.
      required:
        - version
        - operationId
        - kind
        - store
        - recordId
        - payloadDigest
        - predecessor
        - sequence
        - author
        - signature
      properties:
        version:
          type: integer
          enum: [1]
        operationId:
          type: string
        kind:
          type: string
          enum: [put, delete]
        store:
          type: string
        recordId:
          type: string
        payloadDigest:
          type: string
        predecessor:
          type: string
          nullable: true
          description: Digest of the previous mutation. Null if and only if sequence is 0.
        sequence:
          type: integer
          minimum: 0
        author:
          type: object
          required: [identityId, deviceCredential]
          properties:
            identityId:
              type: string
            deviceCredential:
              type: string
        signature:
          type: string
...